# services/sdkapi/docs/Sdk_swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SDK API",
        "title": "Kanthor SDK API",
        "termsOfService": "http://kanthorlabs.com/terms/",
        "contact": {
            "name": "Kanthor Support",
            "url": "http://kanthorlabs.com/support",
            "email": "support@kanthorlabs.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/application/{app_id}/endpoint/{ep_id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointGetRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "endpoint properties",
                        "name": "props",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointUpdateRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointDeleteRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            }
        },
...
        "/endpoint/{ep_id}/rule": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint rule"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maxLength": 32,
                        "minLength": 29,
                        "type": "string",
                        "description": "current query cursor",
                        "name": "_cursor",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "minLength": 2,
                        "type": "string",
                        "description": "search keyword",
                        "name": "_q",
                        "in": "query"
                    },
                    {
                        "maximum": 30,
                        "minimum": 5,
                        "type": "integer",
                        "description": "limit returning records",
                        "name": "_limit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "only return records with selected ids",
                        "name": "_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleListRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint rule"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "rule properties",
                        "name": "props",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleCreateRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            }
        },
        "/endpoint/{ep_id}/rule/{epr_id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint rule"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "rule id",
                        "name": "epr_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleGetRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint rule"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "rule id",
                        "name": "epr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "rule properties",
                        "name": "props",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleUpdateRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint rule"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "endpoint id",
                        "name": "ep_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "rule id",
                        "name": "epr_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointRuleDeleteRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.EndpointRule": {
            "type": "object",
            "properties": {
                "condition_expression": {
                    "description": "examples:\n\t- equal::orders.paid\n\t- regex::.*",
                    "type": "string"
                },
                "condition_source": {
                    "description": "examples\n - app_id\n - type\n - body\n - metadata",
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "endpoint_id": {
                    "type": "string"
                },
                "exclusionary": {
                    "description": "the logic of not-false is true should be used here\nto guarantee default all rule will be on include mode",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "gateway.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "sdkapi.EndpointDeleteRes": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "HTTP: POST/PUT/PATCH",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uri": {
                    "description": "format: scheme \":\" [\"//\" authority] path [\"?\" query] [\"#\" fragment]\nHTTP: https:://httpbin.org/post?app=kanthor.webhook\ngRPC: grpc:://app.kanthorlabs.com",
                    "type": "string"
                }
            }
        },
        "sdkapi.EndpointGetRes": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "HTTP: POST/PUT/PATCH",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uri": {
                    "description": "format: scheme \":\" [\"//\" authority] path [\"?\" query] [\"#\" fragment]\nHTTP: https:://httpbin.org/post?app=kanthor.webhook\ngRPC: grpc:://app.kanthorlabs.com",
                    "type": "string"
                }
            }
        },
...
        "sdkapi.EndpointRuleCreateReq": {
            "type": "object",
            "required": [
                "condition_expression",
                "condition_source",
                "name"
            ],
            "properties": {
                "condition_expression": {
                    "type": "string"
                },
                "condition_source": {
                    "type": "string"
                },
                "exclusionary": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "sdkapi.EndpointRuleCreateRes": {
            "type": "object",
            "properties": {
                "condition_expression": {
                    "description": "examples:\n\t- equal::orders.paid\n\t- regex::.*",
                    "type": "string"
                },
                "condition_source": {
                    "description": "examples\n - app_id\n - type\n - body\n - metadata",
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "endpoint_id": {
                    "type": "string"
                },
                "exclusionary": {
                    "description": "the logic of not-false is true should be used here\nto guarantee default all rule will be on include mode",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "sdkapi.EndpointRuleDeleteRes": {
            "type": "object",
            "properties": {
                "condition_expression": {
                    "description": "examples:\n\t- equal::orders.paid\n\t- regex::.*",
                    "type": "string"
                },
                "condition_source": {
                    "description": "examples\n - app_id\n - type\n - body\n - metadata",
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "endpoint_id": {
                    "type": "string"
                },
                "exclusionary": {
                    "description": "the logic of not-false is true should be used here\nto guarantee default all rule will be on include mode",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "sdkapi.EndpointRuleGetRes": {
            "type": "object",
            "properties": {
                "condition_expression": {
                    "description": "examples:\n\t- equal::orders.paid\n\t- regex::.*",
                    "type": "string"
                },
                "condition_source": {
                    "description": "examples\n - app_id\n - type\n - body\n - metadata",
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "endpoint_id": {
                    "type": "string"
                },
                "exclusionary": {
                    "description": "the logic of not-false is true should be used here\nto guarantee default all rule will be on include mode",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "sdkapi.EndpointRuleListRes": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.EndpointRule"
                    }
                }
            }
        },
        "sdkapi.EndpointRuleUpdateReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "sdkapi.EndpointRuleUpdateRes": {
            "type": "object",
            "properties": {
                "condition_expression": {
                    "description": "examples:\n\t- equal::orders.paid\n\t- regex::.*",
                    "type": "string"
                },
                "condition_source": {
                    "description": "examples\n - app_id\n - type\n - body\n - metadata",
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "endpoint_id": {
                    "type": "string"
                },
                "exclusionary": {
                    "description": "the logic of not-false is true should be used here\nto guarantee default all rule will be on include mode",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "sdkapi.EndpointUpdateReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "sdkapi.EndpointUpdateRes": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "HTTP: POST/PUT/PATCH",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uri": {
                    "description": "format: scheme \":\" [\"//\" authority] path [\"?\" query] [\"#\" fragment]\nHTTP: https:://httpbin.org/post?app=kanthor.webhook\ngRPC: grpc:://app.kanthorlabs.com",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}